Add support for JudgeApps forum feeds.
This is just about ready to be merged. It relies on unreleased
JudgeApps functionality, so I need to land the changes to JudgeApps
before anyone else will be able to use this.

Three types of JudgeApps forum feed: Enter a forum and get all new
threads, enter a forum and get all new posts, or enter a thread and
get all new posts.

TODOs include a user-facing workflow to create these, a user-facing
workflow to attach these to Destinations, formatting work for the
messages, error-handling on the feed fetching, and performance work.

Resolves #7