Allow users to remove destinations from their list
This patch allows users to remove themselves from a destination,
thereby removing it from their list of destinations, in the case of
testing channels or mistakes. This also removes all notifications to
that destination when the last admin is removed, to avoid the case
where there is no one left to configure a destination.

Resolves #36